new blank + move ModInfo definitions to their own file + mod settings are now saved
