+++ loader/src/load/Mod.cpp
@@ -51,22 +51,82 @@ Mod::~Mod() {
     this->unload();
 }
 
-Result<> Mod::loadDataStore() {
-    auto dsPath = this->m_saveDirPath / "ds.json";
+Result<> Mod::loadSettings() {
+    // settings
+
+    // Check if settings exist
+    auto settPath = m_saveDirPath / "settings.json";
+    if (ghc::filesystem::exists(settPath)) {
+        auto settData = file_utils::readString(settPath);
+        if (!settData) return settData;
+        try {
+            // parse settings.json
+            auto data = nlohmann::json::parse(settData.value());
+            JsonChecker checker(data);
+            auto root = checker.root("[settings.json]");
+
+            for (auto& [key, value] : root.items()) {
+                // check if this is a known setting
+                if (auto sett = this->getSetting(key)) {
+                    // load its value
+                    if (!sett->load(value.json())) {
+                        return Err(
+                            "Unable to load value for setting \"" +
+                            key + "\""
+                        );
+                    }
+                } else {
+                    this->logInfo(
+                        "Encountered unknown setting \"" + key + "\" while "
+                        "loading settings",
+                        Severity::Warning
+                    );
+                }
+            }
+
+        } catch(std::exception& e) {
+            return Err(std::string("Unable to parse settings: ") + e.what());
+        }
+    }
+
+    // datastore
+    auto dsPath = m_saveDirPath / "ds.json";
     if (!ghc::filesystem::exists(dsPath)) {
-        this->m_dataStore = this->m_info.m_defaultDataStore;
+        m_dataStore = m_info.m_defaultDataStore;
     } else {
         auto dsData = file_utils::readString(dsPath);
         if (!dsData) return dsData;
-
-        this->m_dataStore = nlohmann::json::parse(dsData.value());
+        try {
+            m_dataStore = nlohmann::json::parse(dsData.value());
+        } catch(std::exception& e) {
+            return Err(std::string("Unable to parse datastore: ") + e.what());
+        }
     }
-    return Ok<>();
+    return Ok();
 }
 
-Result<> Mod::saveDataStore() {
-    auto dsPath = this->m_saveDirPath / "ds.json";
-    return file_utils::writeString(dsPath, m_dataStore.dump(4));
+Result<> Mod::saveSettings() {
+    // settings
+    auto settPath = m_saveDirPath / "settings.json";
+    auto json = nlohmann::json::object();
+    for (auto& [key, sett] : m_info.m_settings) {
+        if (!sett->save(json[key])) {
+            return Err("Unable to save setting \"" + key + "\"");
+        }
+    }
+    auto sw = file_utils::writeString(settPath, json.dump(4));
+    if (!sw) {
+        return sw;
+    }
+
+    // datastore
+    auto dsPath = m_saveDirPath / "ds.json";
+    auto dw = file_utils::writeString(dsPath, m_dataStore.dump(4));
+    if (!dw) {
+        return dw;
+    }
+
+    return Ok();
 }
 
 DataStore Mod::getDataStore() {
@@ -497,239 +557,13 @@ decltype(ModInfo::m_settings) Mod::getSettings() const {
     return m_info.m_settings;
 }
 
+std::shared_ptr<Setting> Mod::getSetting(std::string const& key) const {
+    if (m_info.m_settings.count(key)) {
+        return m_info.m_settings.at(key);
+    }
+    return nullptr;
+}
+
 std::string Mod::getLoadErrorInfo() const {
     return m_loadErrorInfo;
-}
-
-static std::string sanitizeDetailsData(unsigned char* start, unsigned char* end) {
-    // delete CRLF
-    return string_utils::replace(std::string(start, end), "\r", "");
-}
-
-Result<ModInfo> ModInfo::createFromSchemaV010(nlohmann::json const& rawJson) {
-    ModInfo info;
-
-    auto json = rawJson;
-
-    #define PROPAGATE(err) \
-        { auto err__ = err; if (!err__) return Err(err__.error()); }
-
-    JsonChecker checker(json);
-    auto root = checker.root("[mod.json]");
-
-    root.addKnownKey("geode");
-    root.addKnownKey("binary");
-
-    using nlohmann::detail::value_t;
-
-    root
-        .needs("id")
-        .validate(&Mod::validateID)
-        .into(info.m_id);
-    root
-        .needs("version")
-        .validate(&VersionInfo::validate)
-        .intoAs<std::string>(info.m_version);
-    root.needs("name").into(info.m_name);
-    root.needs("developer").into(info.m_developer);
-    root.has("description").into(info.m_description);
-    root.has("repository").into(info.m_repository);
-    root.has("datastore").intoRaw(info.m_defaultDataStore);
-    root.has("toggleable").into(info.m_supportsDisabling);
-    root.has("unloadable").into(info.m_supportsUnloading);
-
-    for (auto& dep : root.has("dependencies").iterate()) {
-        auto obj = dep.obj();
-
-        auto depobj = Dependency {};
-        obj
-            .needs("id")
-            .validate(&Mod::validateID)
-            .into(depobj.m_id);
-        obj
-            .needs("version")
-            .validate(&VersionInfo::validate)
-            .intoAs<std::string>(depobj.m_version);
-        obj
-            .has("required")
-            .into(depobj.m_required);
-        obj.checkUnknownKeys();
-
-        info.m_dependencies.push_back(depobj);
-    }
-
-    for (auto& [key, value] : root.has("settings").items()) {
-        auto sett = Setting::parse(key, value.json());
-        PROPAGATE(sett);
-        info.m_settings.insert({ key, sett.value() });
-    }
-
-    if (auto resources = root.has("resources").obj()) {
-        for (auto& [key, _] : resources.has("spritesheets").items()) {
-            info.m_spritesheets.push_back(info.m_id + "/" + key);
-        }
-    }
-
-    root.has("binary").asOneOf<value_t::string, value_t::object>();
-
-    bool autoEndBinaryName = true;
-
-    root.has("binary").is<value_t::string>().into(info.m_binaryName);
-    
-    if (auto bin = root.has("binary").is<value_t::object>().obj()) {
-        bin.has("*").into(info.m_binaryName);
-        bin.has("auto").into(autoEndBinaryName);
-
-    #if defined(GEODE_IS_WINDOWS)
-        bin.has("windows").into(info.m_binaryName);
-    #elif defined(GEODE_IS_MACOS)
-        bin.has("macos").into(info.m_binaryName);
-    #elif defined(GEODE_IS_ANDROID)
-        bin.has("android").into(info.m_binaryName);
-    #elif defined(GEODE_IS_IOS)
-        bin.has("ios").into(info.m_binaryName);
-    #endif
-    }
-
-    if (
-        root.has("binary") &&
-        autoEndBinaryName &&
-        !string_utils::endsWith(info.m_binaryName, GEODE_PLATFORM_EXTENSION)
-    ) {
-        info.m_binaryName += GEODE_PLATFORM_EXTENSION;
-    }
-
-    if (checker.isError()) {
-        return Err(checker.getError());
-    }
-    root.checkUnknownKeys();
-
-    return Ok(info);
-}
-
-Result<ModInfo> ModInfo::create(nlohmann::json const& json) {
-    // Check mod.json target version
-    auto schema = LOADER_VERSION;
-    if (json.contains("geode") && json["geode"].is_string()) {
-        auto ver = json["geode"];
-        if (VersionInfo::validate(ver)) {
-            schema = VersionInfo(ver);
-        } else {
-            return Err(
-                "[mod.json] has no target loader version "
-                "specified, or it is invalidally formatted (required: \"[v]X.X.X\")!"
-            );
-        }
-    } else {
-        return Err(
-            "[mod.json] has no target loader version "
-            "specified, or it is invalidally formatted (required: \"[v]X.X.X\")!"
-        );
-    }
-    if (schema < Loader::s_supportedVersionMin) {
-        return Err(
-            "[mod.json] is built for an older version (" + 
-            schema.toString() + ") of Geode (current: " + 
-            Loader::s_supportedVersionMin.toString() +
-            "). Please update the mod to the latest version, "
-            "and if the problem persists, contact the developer "
-            "to update it."
-        );
-    }
-    if (schema > Loader::s_supportedVersionMax) {
-        return Err(
-            "[mod.json] is built for a newer version (" + 
-            schema.toString() + ") of Geode (current: " +
-            Loader::s_supportedVersionMax.toString() +
-            "). You need to update Geode in order to use "
-            "this mod."
-        );
-    }
-
-    // Handle mod.json data based on target
-    if (schema <= VersionInfo(0, 2, 0)) {
-        return ModInfo::createFromSchemaV010(json);
-    }
-
-    return Err(
-        "[mod.json] targets a version (" +
-        schema.toString() + ") that isn't "
-        "supported by this version (v" + 
-        LOADER_VERSION_STR + ") of geode. "
-        "This is probably a bug; report it to "
-        "the Geode Development Team."
-    );
-}
-
-Result<ModInfo> ModInfo::createFromFile(ghc::filesystem::path const& path) {
-    try {
-        auto read = file_utils::readString(path);
-        if (!read) return Err(read.error());
-        auto res = ModInfo::create(nlohmann::json::parse(read.value()));
-        if (!res) return res;
-        auto info = res.value();
-        info.m_path = path;
-        return Ok(info);
-    } catch(std::exception const& e) {
-        return Err(e.what());
-    }
-}
-
-Result<ModInfo> ModInfo::createFromGeodeFile(ghc::filesystem::path const& path) {
-    ZipFile unzip(path.string());
-    if (!unzip.isLoaded()) {
-        return Err<>("\"" + path.string() + "\": Unable to unzip");
-    }
-    // Check if mod.json exists in zip
-    if (!unzip.fileExists("mod.json")) {
-        return Err<>("\"" + path.string() + "\" is missing mod.json");
-    }
-    // Read mod.json & parse if possible
-    unsigned long readSize = 0;
-    auto read = unzip.getFileData("mod.json", &readSize);
-    if (!read || !readSize) {
-        return Err("\"" + path.string() + "\": Unable to read mod.json");
-    }
-    nlohmann::json json;
-    try {
-        json = nlohmann::json::parse(std::string(read, read + readSize));
-    } catch(std::exception const& e) {
-        delete[] read;
-        return Err<>(e.what());
-    }
-
-    delete[] read;
-
-    if (!json.is_object()) {
-        return Err(
-            "\"" + path.string() + "/mod.json\" does not have an "
-            "object at root despite expected"
-        );
-    }
-
-    auto res = ModInfo::create(json);
-    if (!res) {
-        return Err("\"" + path.string() + "\" - " + res.error());
-    }
-    auto info = res.value();
-    info.m_path = path;
-    
-    // unzip known MD files
-    using God = std::initializer_list<std::pair<std::string, std::string*>>;
-    for (auto [file, target] : God {
-        { "about.md", &info.m_details },
-        { "changelog.md", &info.m_changelog },
-    }) {
-        if (unzip.fileExists(file)) {
-            unsigned long readSize = 0;
-            auto fileData = unzip.getFileData(file, &readSize);
-            if (!fileData || !readSize) {
-                return Err("Unable to read \"" + path.string() + "\"/" + file);
-            } else {
-                *target = sanitizeDetailsData(fileData, fileData + readSize);
-            }
-        }
-    }
-
-    return Ok(info);
 }
